Major restructuring of the FProject code that is not using sourcePlane from the main VKL code anymore. The templates for creating mocks have been updated. Documentation has been included with all the different options, but creating the actual json file still needs to be tested.

+++ analysis_tools/finalize.sh
@@ -119,7 +119,10 @@ if [ "$lmodel" = "pert" ]; then
     cp ${path}${run}output/smooth_residual.fits $target/smooth_residual.fits
 fi
 
-rm -r ${path}${run}analysis
+
+if [ -d ${path}${run}analysis ]; then
+    rm -r ${path}${run}analysis
+fi
 cp -r $target/ ${path}${run}analysis
